I improved my modules a lot, list of modules: Diary_Slim, Tasks, Watch_History, GamePlayer, Stories, and Module_Selector
I significantly improved the "Write_On_Diary_Slim" subclass of the "Diary_Slim" class, removing fixed and language-dependent text.
I also improved the state texts-related methods and variables on the root class.
And removed additional text dictionaries for the Slim texts dictionary, the text now uses the "Slim texts.json" file to define its variables.

I converted the "Tasks" module to the new style with the new classes, removing the old functions.
This finishes my conversion of all my Python modules to a new style with the new classes, not using the old unorganized functions anymore.
I have 16 (sixteen) modules that are in the new style.

In the "Watch_History" module, I added the possibility to type "zero" as the number of watched times for Re-Watching media.
If the user types zero, then the Re-Watching status is deactivated for the media, only for the current media item (episode, video, or movie).
I also made the "Watch_Media" class check if today is Christmas using the "Today_Is_Christmas" method of the Christmas class.
If today is Christmas and the media is series media (not a movie) and not a video, it replaces the "Episode" text with the "Christmas special episode" text.
It replaces the "Episode" text on the "media_unit_name" key on the media dictionary.
To be shown using the "Show_Media_Information" method of the "Watch_History" class.

In the "GamePlayer" module, I fixed some errors in the "Register_Playing_Time" class.
It was not registering the language game playing times on the "game played texts" folder, I created a method for that.
The game playing time text that is written on the game text file was also not right.
I fixed that, removing the part of the game playing time that shows the game category and name.
I fixed an error on the "Time_Text" method of the "Date" class and on the "Play" class of the "GamePlayer" module.
That error was causing the hour text to not be added to the game playing time text.
I created the "Game names" JSON file to contain game names that can not be defined on a shortcut file, such as game names with colons, question marks, and quotes.

In the "Stories" module, I updated it to add the "Chapter number" key to each story information dictionary.
The chapter number also exists in the "Information.json" file in the information folder of the story.
It is updated using the length of the "chapter titles" list in English (to get the chapter number).

I improved the "MS" (Module_Selector) Python module a lot, removing the old way of checking command-line arguments.
And I created the "Main" class to check arguments, and switches, and also let the user select one class to run if it did not type any arguments.
It also resets the switches at the end of the execution of the class, the switches are now in the JSON format for easier changing of keys.