In the "Food_Time" class, I created the "Set_Timer" method to open a countdown to the "will be hungry" time on the "Time And Date" (timeanddate.com) website. And to schedule a task on the operating system to play an alarm sound file on the browser when the "will be hungry" time is reached. I also created the "Play_Alarm" class to be used on the path of the scheduled task, to open the alarm sound file on the browser.
I renamed the "App Text Files" folder to "Module Files".
I also modified all of the "Define_Module_Folder" methods of every class of every non-utility module to make the definition of the contents of the "Module Files" and "Modules" folders dynamic, using the "Contents" method of the "Folder" class.