I renamed the "App Text Files" folder to "Module Files". I also modified all of the "Define_Module_Folder" methods of every class of every non-utility module to make the definition of the contents of the "Module Files" and "Modules" folders dynamic, using the "Contents" method of the "Folder" class.
In the "Unblock" class of the "Block_Websites" module, I added the ability to let the user create a website list to unlock by selecting websites from the websites list.
In the "Block" class of the same module, I updated the text that the class writes to the "hosts" file, adding a header showing the "Block_Websites.py" text.