In the "Folder" class, I improved the creation of the folders on the folders dictionary, making it more dynamic with "for" loops, and also added the ability to create files using the "File" class.
In the "Tasks" class, I improved the definition of its folders, using the "Contents" method of the "Folder" class.
I also created a new "Tasks.json" file and added more things to be written inside it.
Like the number of total tasks, the list of task types, task names in English and Portuguese, and task descriptions in English and Portuguese.
The "Tasks.json" file also exists inside all of the "Per Task Type" task type folders, but with only the task type dictionary.
I improved the "Tasks" JSON dictionary a lot, adding "ISO8601" and language times, and this file name format: "Number. Task Type (Time)".
I remade the "Register_Task" class to write on the new "Tasks.json" file and the same files on the tasks type folders.
I also added a "Task list.txt" text file on the root folder and on each task type folder to contain the file names of the tasks.
And a "Files" folder on each task type folder contains the task files, with a text containing all of the information about the task.
I also created a new class called "Show_Tasks" to show all the information about the registered tasks.

In the "Date" class, I improved the keys of the dictionary of the "Now" module, adding more date, time, and datetime formats.
I also added day name, month name, date, and datetime formats in each language, as the "day_names", "month_names", "date_format", and "datetime_format" dictionaries.