I started to remake the code of the "Watch_History" module to have the information about the media and media types in JSON format. And also to register the watched episodes or movies in a file called "Episodes.json" file on the "root watched folder" and "media type watched folder". I remade all of the methods on the root class to improve them, and to use the same "media dictionary" for selecting a media type and a media. I also integrated the code of the "Define_Media_Variables" method of the "Watch_Media" class into the "Define_Media_Dictionary" method.
I improved the main class of the "Module_Selector" or "MS" module a lot, improving the definition of the argument parser.
I moved the code related to the argument parser to the new method called "Define_Parser" to make the code more organized.
I added dictionaries for default arguments, options, and function arguments to pass to ArgumentParser.
I created the "Add_Argument" method to add arguments and replaced the original "help" text with a new help text with a description and epilogue.
I also a text explaining each of the "testing", "verbose", and "user information" arguments in the user language.
I added the ability for the modules to define a list of argument names to add to the list of arguments that can be used to run the module.
And the ability for the modules to add a list of additional arguments to be added to the arguments list, using the specified explanatory text in the argument dictionary.
Like the "Food_Time" module that has the additional arguments "set" and "check", the "set" one is for defining the meal time, and the "check" is for displaying it.