On the "Watch_History" Python module, I made the following changes: In the "Manage" subclass, in the "Convert_History" subclass, I created the new "Fix_Comment_Dictionaries" method. It was used by me to fix some Comment dictionaries inside watched media Entry dictionaries. Which had duplicate keys, dates equal to the entry date, or only Entry or Date keys.
On the "Stories" Python module, I made the following changes:
I added the new "Theme" dictionary to the "Pack" dictionary of all stories, and began adding the story colors to the Theme dictionaries.
I also added the new "Parent story" system, updating the old "Website link name" key on all stories that have a parent story.
Now stories can define a parent story, to be used for changing the website link of the story.
The story websites that have a parent story can better define their parent story, with a title, and a custom folder.
I also made improvements on the definition of the "Pack", "Theme", "Parent story", "Chapter dates, and "Authors" dictionaries and lists.