On the "Module_Selector" module, I improved a lot on the definition of alternate and custom arguments for the modules. The help text for those arguments will now be inside the "Texts" file of the specific module, and not inside the "Texts" file of "Module_Selector". Thus making those arguments less dependent on the "Module_Selector". The alternate arguments are alternative names for the module to be used in the command line. For example, the "GamePlayer" module has an argument with its name, and also the "-play" alternate argument. Now the modules can define custom arguments which are dictionaries, with custom actions and keys. Each argument of that type will now have the title of the module appended to the beginning of the argument key, along with a dot. I did that to not use arguments that are common words like "set" or "check", that sometimes may be used as module titles. Modules that have a title that is a common word are the "Code", "Diary", "Friends", "Stories", "Tasks", and "Years" modules. For example, the "Thing" module has a "make" argument, so the argument key would be "thing.make", and not just "make". Custom arguments are additional arguments that are added alongside the argument of the module title. Their purpose is to store valuable information that will be used by the module. Information such as a class to be auto-executed, or a text or number to be used to auto-select options in options lists. I also documented and added a lot of comments on the code of "Module_Selector", for a better understanding of various parts of the code.
On the "Food_Time" module, I updated its code to use the new implementation of the "custom arguments" of the "Module_Selector" module.
Now it parses the custom arguments using its new root method called "Parse_Arguments".
If no argument is present, then the meal times will be registered.
If the "set" argument is active, then the meal times will also be registered.
If the "check" argument is active, then the meal times will be shown on the screen.

On the "GamePlayer" module, I added the new "game" custom argument.
It has the auto-class "Play", to automatically run it when the argument is present and has a value.
The value of the argument is where the game title is stored, to automatically select it to play.