On the "Friends" and "Social_Networks" modules, I improved them a lot. I greatly improved the definition and generation of the "Friends" and "Social Networks" dictionaries. I also improved the definition of the information items of each class, sharing similar code between the classes. And making more efficient definitions of the information items of both friends and social networks.
On the "Friends" module, on its "Open_Friend_File" subclass, I fixed and greatly improved it.
I improved the way of searching for friends using information items, the states, and the visual aspect of the subclass.
It can search using all of the information items of all friends or do not use a search and only select a friend to open its file.
It can also search through all information items of all social networks.
Finding a social network of a friend that matches the searched query.
And opening the file of the found friend.