On the "Folder" module, I improved the definition of the folder dictionary a lot. Improving keys, adding new keys that look better, and folder dictionaries with keys and English and folder names in the user language.
On the "Years" module, I improved and fixed its variables, methods, and sub-classes a lot.
I remade its "Years" dictionary using the dictionaries of the "Social_Networks" and "Friends" modules as a base.
I also fixed and improved the "Create_Year_Summary" and "Verify_Current_Year" sub-classes.
I improved the definition of year folders, using the definition of folders of the "Social_Networks" module as a base.

I improved and fixed the "Christmas" module a lot, improving its variables and dictionaries.
The style of showing information to the user, and the texts.
I also improved the opening of Social Networks and added more steps to the Christmas planning.