In the "Stories" module, I have made the following changes: In the root class, I greatly improved the definition of the story links dictionary. This now contains the keys "Website", "Wattpad", and "Spirit Fanfics". The class defines the story's website links, and the story IDs, links, and custom links on Wattpad and Spirit Fanfics. Now the "Website" dictionary is written to its new JSON file of the same name. The same goes for the "Spirit Fanfics" links dictionary and its new JSON file.
In the root class, I have made the following changes:
I created the new root method "Define_Story_Websites_Dictionary".
To define the dictionaries of story websites, with their link templates and additional links.
It also creates the "Story websites" dictionary, writing on the JSON file of the same name.
And I also created the "Define_Story_Websites" method.
To define the story website dictionary inside each story dictionary.

In its "Create_New_Story" sub-class, I added more requests for information about the story.
For example, ask the user if the story has a parent story, if so, ask for the title of the parent story.
The sub-class also asks if the user posted the story on the story "Wattpad" and "Spirit Fanfics" story websites.
If so, the sub-class asks for the IDs of the story on both websites and in all supported languages.
Thus creating the story website dictionaries.
I also moved the story dictionary key definition code to the new "Define_Story_Dictionary" method, to make it more organized.