In the "Diary_Slim" module, I added the new methods "Define_Statistics", "Update_Statistics", "Update_Current_Year_Statistics", and "Verify_Input". The "Define_Statistics" method defines the general statistics dictionary and for each year. These statistics include things that I did during the year, like bike rides, driving, showers taken, times I washed my clothes, drawings, posted videos, work done, and family visits. It creates these dictionaries with the data and writes the general statistics dictionary to the "Statistics.json" file, which is located in the data folder of the Diary Slim. The method also defines a dictionary of years to store the folders and statistics dictionaries for each year. After that, it iterates through the text dictionary of Diary Slim to find texts that contain the key "Add to statistics". These texts have a list of custom statistics that will be added. Then, the "Define_Statistics" method calls the "Update_Statistics" method to update the general statistics file, for all years, and update the "Current year" dictionary of Diary Slim, which will be used later. The "Update_Current_Year_Statistics" method receives a modified statistics dictionary as a parameter, it updates the statistics file of the current year using that dictionary. The "Verify_Input" method verifies what the user types. It defines the input text to be shown to the user in their language. Furthermore, it checks if the requested input needs to be an integer. If so, it uses a regular expression to ensure that the input is indeed an integer. Finally, I updated the "Define_Additional_Information" method to make it more modular. This helps in defining statistics with responses of "Yes or no", "Type", and "Select".
The statistics have a main key, which is the name of the statistic in English.
For example, the Diary Slim text "Bike" corresponds to the key "Bike rides".
They can include questions, like "How many times did you drive the car?", and the expected type of response, which can be "Yes or no", "Type", or "Select".
Also, there is a dictionary of responses, in the case of "Yes or no", if the user answers "yes", it uses the "yes" dictionary that is inside the "Answers" dictionary.
For instance, if the question is "Did you create new websites?", and the answer is yes, it says "Type how many websites you create".
The statistics can also have additional questions, which are stored in a dictionary called "Questions".
For example, in the case of in-person work, the main question is "What is the name of the workplace?", and an additional question is "How much money did you earn?".
This way, the statistics can have multiple additional questions, each with its own statistic numbers.

In the "Write_On_Diary_Slim" sub-class, I modified the "Define_Item_Variables" method to include the option for alternative text formatting.
Normally, it is the root method "Define_Additional_Information" who formats the texts, in a simple way, just replacing the original text with a new formatted one.
But I needed something more flexible, like the ability to select items from a list to format, as well as using texts with different genders, based on a list that contains them.
For example, I have a text in Diary Slim that I use to register when my siblings come over to my house.
So, I select which sibling visited, and the method checks the gender of that sibling, based on two lists: one with male names, and another with female names.
Then, it formats the text that will be written in the Diary Slim, inserting the name of the sibling, and using "My brother" or "My sister", depending on the gender.

In the same sub-class, I created a new method called "Manage_Statistic" (Gerenciar_Estatística).
This method is executed when the Diary Slim text is also a statistic, and it retrieves the corresponding dictionary for that statistic.
If the statistic has a question, the method calls "Define_Additional_Information" to ask the user for the necessary information.
The responses can be "yes", followed by a question, "no", typing something, or selecting an option from a list.
Next, the method defines a statistic text in the user's language, which will be shown later, considering some conditions that may change which text will be used.
It also sets a key that will be used to locate the statistic in the current year's statistics dictionary.
The method creates a statistics dictionary within the "Text" dictionary of the root dictionary, containing the number of statistics to be updated.
Also, it includes the key "Dictionary", which contains a dictionary of the root statistic, with the text in the user's language, the old number, and the new number, which will still be changed.
This dictionary also has the key "Changed statistic", which has a default value of "False".
Then, the method sets the number to be added as one and checks if the statistic contains a question.
If it does, it tries to get the answer from the information dictionary, if the answer is a digit, it converts it into a number.
If the answer is "no" or zero, the number to be added is also set to zero.
If the key "Questions" is present in the statistic dictionary, the method iterates through the list of questions, sets the old number, and adds the root number to the current year's statistic number.
It also adds the additional statistics to the Diary Slim statistics dictionary, including the text, the old number, and the current number.
After that, the method counts how many statistics need to be changed and increases the statistic number using the amount of the number to be added.
If the old number and the new number are different, the method changes the state of "Changed statistic" to "True".
Then, it checks if there is a text decorator in the Diary Slim text dictionary.
If there is, it modifies the text that will be shown to the user, using the defined text decorator.
For example, if the normal text is "[Name of the Person]" and the text decorator is "Visits of", the text with decorator would become "Visits of [Name of the Person]."
This is a way to present the statistic text in a more appealing manner for the user.
After that, the method updates the statistics for the current year, using the local statistics dictionary that has already been modified.

The "Write" method checks if the Diary Slim text is a statistic and if it has been altered.
If both conditions are true, it sets a text based on the number: "Changed statistic" for a single statistic and "Changed statistics" for two or more.
Then, it displays this text along with the updated statistics.
The text for each statistic appears in the user's language, including " in [Current year]", the current number of the statistic, and " (before: [old number])".