In the "Tasks" module, I have made the following changes: Now the "Register" class can also register a task's progress, and not just a full task. This way, the user can register their progress on tasks in the current Diary Slim without needing to register a task. I did this by creating a new state called "Register task" that can be changed by the "Diary Slim" module. This state determines if the task will be registered in the database of productive things done (Task History). Or if the task will only be registered in the current Diary Slim, with the class writing the task description there. The task types can also use different task titles for the task progress. For example, the normal title of the Python task type is: "I programmed in Python a lot, programming my module [X]". And the title for the task progress is: "I have been programming my modules in Python until now". I also updated how the root method "Show_Information" shows information about the registered task to the user when it is a task progress. Changing "This task was registered" to "The progress of this task was registered", and changing other texts to talk about the task progress rather than the task itself.
In the "Diary_Slim", in the "Write_On_Diary_Slim" class, I have made the following changes:
I added the "Register task" state to the task dictionary (when the selected text is also a task).
After defining the task dictionary with that state, the class asks if the user wants to register the full task or just the progress of the task.
If the user wants to register only the progress of the task, then the "Register task" key will be set to False (No).
I also changed various conditions throughout the class to check if the task was registered or not.
If the text is also a task and has a statistic, even if the task was or was not registered, the root number of the statistic will go up by one.
But the statistic questions will only be asked if the task was registered, not if it is just a task progress.
For example, in the Python task type, the statistic "Times programmed in Python" will go up by one even if a task progress is registered.
And in the PHP task type, the statistic "Times programmed in PHP, HTML, CSS, and JavaScript" will also go up by one, but the sub-statistic "New websites created" will not.
This is because this sub-statistic depends on a question, and questions are only asked when full tasks are registered, not just a task progress.