2025-10-25 - In the "Folder" utility module, I moved the definition of the "links" and "website" dictionaries to the "PHP" usage module. In the root class of the "PHP" module, in the "Define_Dictionaries" method, the "PHP" class now creates the new root "website" dictionary, which is the fusion of the two previous dictionaries. This dictionary contains an URL format in the "URL format" key, and two empty keys to store the domain and sub-domain of the website. Then the method reads the "Website.json" file that is in the folder where my websites are, to get the two items mentioned last. After that, it creates the "URL" key using the found domain and sub-domain. And then writes the URL and URL format keys in the "Website.json". An important PHP file named "Website.php" also has the same functionality of creating the URL and writing it into the "Website.json" file, but written in PHP code, so that PHP is independent of Python.
In the "JSON" utility module, I created the new "Remove_Duplicates_From_List" method to remove duplicate items from a list.

In the "Stories" module, I changed its root class to create the new "Story links.txt" and "Links.json" files inside the story information folder.
The "Story links.txt" file contains the story website links, both the root link and the language links (Portuguese and English).
It also contains the story links in additional story websites like "Wattpad" and "Spirit Fanfics", if the story was posted on those websites.
While the "Links.json" file contains the story links dictionary in JSON format, with the story website links and the story links in the additional story websites, if they exist.
I also changed the root class to create the new "Story titles.txt" and "History of story titles.txt" files inside the story information folder.
The "Story titles.txt" file is filled by the root class with the story titles in the order of the languages of the "Language" utility module.
While the "History of story titles.txt" file remains empty and can be manually filled by the user, since I do not have a history that registers the change in the story titles.
I also made various code and comment improvements in the root class and its methods.
I started modifying the chapters dictionary of the stories to change it to the new format that I developed.
In the new format, the chapter titles are kept inside the "Lists" dictionary.
The chapter writing dates list in the "Writing dates" key was removed because it is not necessary anymore.
There is a new "List" key to list the chapter numbers from one to the last chapter written.
I still have not finished changing the way the classes of the "Stories" module access the chapters dictionary for each story.
I still need to change that in the root class and in the "Create_New_Story", "Post", and "Write" sub-classes.