2025-10-29 - In the "JSON" utility module, in its root class, I made the following changes: I created the new root method "List_Has_Only_Numbers" to determine if a list contains only numbers. I changed its root method "Sort_Item_List" to sort the keys of a dictionary or the items of a list based on a list with orders that are numbers.
In the "Stories" module, in its root class, I made the following changes:
Now the definition of the story websites dictionary in the root method "Define_Story_Websites_Dictionary" is dynamic, no longer dependent on predefined story websites like "Wattpad" and "Spirit Fanfics".
I created a new folder called "Story Websites" inside the story database folder.
In this folder, there is a folder for each story website, and inside each website folder, there is a file called "Website.json".
This JSON file contains the name of the story website, the order in which it should appear, and a dictionary of the website's links.
The links contain link templates such as "{Website Link}" and "{Profile Link}" that should be replaced by the respective link in the method that defines the story website dictionary.
And after the template, there are parts of the link that define the pages of a story, such as the main story page, a page to edit the story, manage its chapters, or add a new chapter.
I changed the root method "Define_Story_Websites" that defines the story websites for each story.
I did this to improve how it creates the links for each story to use the new format.
Now the story links on Wattpad no longer include the story title, only the story ID.
I removed the old "Define_Social_Network_Variables" root method because it defined the links for the social networks and story websites "Wattpad" and "Spirit Fanfics" in a non-dynamic way and was no longer needed.
I changed the root method "Define_Stories_Dictionary" to list the stories from their folders, and not from a predefined list in the "Stories.json" file.
Now it lists all the folders inside the stories folder, removes some non-story folders based on a predefined list in the method and also from a list in the new "Ignored Folders.txt" file.
After that, it iterates over all the story folders, reads their "Story.json" files, gets the story titles, and adds them to the "List" key and the "Titles" dictionary.
I also removed the "Language" and "All" keys from the "Titles" dictionary because they were no longer needed.