2025-11-01 - In the "Language" utility module, in its root class, I imported the "Create", "Open", "Content", and "Edit" methods from the "File" class to handle files, as the methods in the "File" class were prettier and had better comments. I also imported the "Edit", "Dumps", "From Python", and "To Python" methods from the "JSON" class to handle JSON files and dictionaries, converting "From Python" to JSON, and from JSON "To Python", for the same reasons mentioned in the previous line. In the "Define_Languages" method, I created the new language dictionary that contains dictionaries for each language. Each dictionary contains the keys "Small", "Full", and "Translated", referring to the language in question. The dictionaries receive the language key in small format. After that, the method updates the "Languages.json" file to add the updated language dictionary. I removed methods from the "Language" class that were not being used and also improved the code comments throughout the class. In the usage modules that use full or translated languages, I changed their usage to iterate over the new languages dictionary, to use a more standardized approach.
In the "Stories" module, I made the following changes:
In its "Write" sub-class, I renamed the root dictionary "dictionary" to "writing" to better reflect its purpose.
I merged the "Define_Steps" and "Run_Steps" methods into a single method called "Define_And_Run_Steps", which now defines and executes the chapter writing steps.
Each step is associated with a method in the "Write" sub-class.
I significantly improved the "Select_Writing_Mode" method.
I improved the definition of each writing mode, now using the new "Finished" status key from the chapter writing dictionaries to check if the chapter has already been written, and if so, advance to the next chapter.
When advancing to the next chapter, the "Current Chapter" key is updated, the "Times" dictionary keys are set to empty, and the "Finished" key is changed to no.
If the user has already revised all chapters, they can now select any chapter to revise, and the "Select Chapter" state is activated.
In translation mode, if all chapters have already been translated, the method disables the translation mode.
And if the current chapter to be translated is the same one to be revised, and the chapter has not yet been revised, the method also disables the translation mode, and it can no longer be selected by the user.
The new "Ongoing writing session" state determines whether that is an ongoing writing session, and the default is no.
If the "Started" writing time is not empty, the state becomes yes, as it is an ongoing writing session.
The method now creates custom texts for each writing mode based on the chapter's current status.
"Start to [write/revise/translate]" for chapters that are yet to be [written/revised/translated].
"Continue to [write/revise/translate]" for chapters that are currently being [written/revised/translated].
I also significantly improved the "Define_Chapter" method.
Now, if the "Select chapter" state is activated, it executes the new "Select_Chapter" method, which prompts the user to select a chapter to revise.
I redesigned the chapter dictionary to make its keys more organized and with nicer names, and I also removed keys that were no longer necessary.
I added the new "Writing language" dictionary as requested in the programming plan for the "Stories" module, to store the language in which the chapter will be written or revised.
I also added the new "Origin language" and "Destiny language" language dictionaries to store the original language of the chapter and the language in which it will be translated, all for translation mode.
Now the "Define_Chapter" method opens the chapter file only in the user's language if the writing mode is writing or revision.
But it still opens the chapter files in all languages if the writing mode is translation.
I improved the "Check_Chapter_Date_Texts" method to update the chapter date texts to a better version.
And also to access the chapter's writing dictionary in the new format I created.
Using the chapter's writing date for the writing mode, and using the last revision and translation dates, all to update the chapter dates in its file.

I started programming the code for the section about the "Language" module and the entire section of the "Stories" module at 02:55 on November 1, 2025.
And I stopped programming at 12:48 on the same day.
During that time, I programmed for 5 hours, 39 minutes, and 57 seconds (05:39:57), including breaks.